+++ src/components/artifacts/ArtifactPopup.svelte
