+++ src/components/artifacts/ArtifactSlotWrapper.svelte
@@ -2,7 +2,7 @@
     import type { CharacterArtifactData } from "@/types";
     import { artifactSlots } from "@/constants";
     import ArtifactSlot from "./ArtifactSlot.svelte";
-    import { userArtifactStore } from "../lib/storage";
+    import { userArtifactStore } from "../../lib/storage";
 
     interface Props {
         characterName: string;
